refactor: metrics generation to metrics plugin
