--- internal/metric/prometheus/prometheus.go
@@ -1,282 +1,0 @@
-/**
- * Copyright (c) F5, Inc.
- *
- * This source code is licensed under the Apache License, Version 2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-package prometheus
-
-import (
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"log"
-	"log/slog"
-	"net/http"
-	"os"
-	"strconv"
-	"strings"
-	"time"
-
-	metrics "github.com/nginx/agent/v3/internal/metric"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/metric"
-	"go.opentelemetry.io/otel/sdk/metric/metricdata"
-
-	metricSdk "go.opentelemetry.io/otel/sdk/metric"
-)
-
-const (
-	defaultTarget  = "http://127.0.0.1:30882/metrics"
-	DataSourceType = "PROMETHEUS"
-	meterName      = "nginx-agent-prometheus-scraper"
-)
-
-// Used for deserializion of parsed Prometheus data.
-type (
-	// A single data point for an entry. An entry can have multiple points.
-	DataPoint struct {
-		Name   string
-		Labels map[string]string
-		Value  float64
-	}
-
-	// Represents a single entry for a Prometheus meter, is one of [counter, gauge, histogram, summary].
-	DataEntry struct {
-		Name        string
-		Type        string
-		Description string
-		Values      []DataPoint
-	}
-
-	Scraper struct {
-		meter                       metric.Meter
-		producer                    *metrics.MetricsProducer
-		previousCounterMetricValues map[string]DataPoint
-		cancelFunc                  context.CancelFunc
-	}
-)
-
-func NewScraper(meterProvider metric.MeterProvider, producer *metrics.MetricsProducer) *Scraper {
-	meter := meterProvider.Meter(meterName, metric.WithInstrumentationVersion("v0.1"))
-	return &Scraper{
-		meter:                       meter,
-		producer:                    producer,
-		previousCounterMetricValues: map[string]DataPoint{},
-	}
-}
-
-func (s *Scraper) Start(ctx context.Context, updateInterval time.Duration) error {
-	slog.Info("prometheus data source booting up")
-	ticker := time.NewTicker(updateInterval)
-
-	cancellableCtx, cancelFunc := context.WithCancel(ctx)
-	s.cancelFunc = cancelFunc
-
-	// This should be discovered dynamically.
-	targets := []string{defaultTarget}
-	if value, ok := os.LookupEnv("PROMETHEUS_TARGETS"); ok {
-		targets = strings.Split(value, ",")
-	}
-
-	for _, target := range targets {
-		resp, err := http.DefaultClient.Get(strings.TrimSpace(target))
-		if err != nil {
-			return fmt.Errorf("could not find prometheus endpoint %s: %w", target, err)
-		}
-
-		log.Printf("found prometheus endpoint %v", resp.Request.URL)
-	}
-
-	for {
-		select {
-		case <-cancellableCtx.Done():
-			return nil
-		case <-ticker.C:
-			for _, target := range targets {
-				entries := scrapeEndpoint(strings.TrimSpace(target))
-
-				for _, entry := range entries {
-					switch entry.Type {
-					case "gauge":
-						s.addGauge(entry)
-					case "counter":
-						s.addCounter(entry)
-					case "histogram":
-						s.addHistogram(entry)
-					case "summary":
-						slog.Debug("cannot report 'summary' metric: unsupported")
-					}
-				}
-			}
-		}
-	}
-}
-
-func (s *Scraper) Stop() {
-	slog.Info("prometheus data source shutting down")
-	s.cancelFunc()
-}
-
-func (s *Scraper) Type() string {
-	return DataSourceType
-}
-
-func (s *Scraper) addGauge(de DataEntry) {
-	gauge := metrics.NewFloat64Gauge()
-
-	for _, point := range de.Values {
-		_, err := s.meter.Float64ObservableGauge(
-			de.Name,
-			metric.WithDescription(de.Description),
-			metric.WithFloat64Callback(gauge.Callback),
-		)
-		if err != nil {
-			slog.Error("failed to initialize OTel gauge from Prometheus data", "error", err)
-		}
-
-		metricAttributes := []attribute.KeyValue{}
-		if len(point.Labels) > 0 {
-			for labelKey, labelValue := range point.Labels {
-				metricAttributes = append(metricAttributes, attribute.KeyValue{Key: attribute.Key(labelKey), Value: attribute.StringValue(labelValue)})
-			}
-		}
-		gauge.Set(point.Value, attribute.NewSet(metricAttributes...))
-	}
-}
-
-func (s *Scraper) addCounter(de DataEntry) {
-	counter, _ := s.meter.Float64Counter(
-		de.Name,
-		metric.WithDescription(de.Description),
-	)
-
-	for _, point := range de.Values {
-		metricAttributes := []attribute.KeyValue{}
-		for labelKey, labelValue := range point.Labels {
-			metricAttributes = append(metricAttributes, attribute.KeyValue{Key: attribute.Key(labelKey), Value: attribute.StringValue(labelValue)})
-		}
-
-		if previousMetricValue, ok := s.previousCounterMetricValues[de.Name]; ok && point.Value > 0 {
-			counter.Add(context.TODO(), point.Value-previousMetricValue.Value, metric.WithAttributeSet(attribute.NewSet(metricAttributes...)))
-		} else {
-			counter.Add(context.TODO(), point.Value, metric.WithAttributeSet(attribute.NewSet(metricAttributes...)))
-		}
-
-		s.previousCounterMetricValues[de.Name] = point
-	}
-}
-
-func (s *Scraper) addHistogram(de DataEntry) {
-	histogram := metricdata.HistogramDataPoint[float64]{
-		Bounds:       []float64{},
-		BucketCounts: []uint64{},
-		StartTime:    time.Now(),
-		Time:         time.Now(),
-	}
-
-	for _, point := range de.Values {
-		var bound string
-		metricAttributes := []attribute.KeyValue{}
-		for labelKey, labelValue := range point.Labels {
-			if labelKey == "le" {
-				bound = labelValue
-			}
-			metricAttributes = append(metricAttributes, attribute.KeyValue{Key: attribute.Key(labelKey), Value: attribute.StringValue(labelValue)})
-		}
-
-		if strings.HasSuffix(point.Name, "_bucket") && bound != "" {
-			histogram.BucketCounts = append(histogram.BucketCounts, uint64(point.Value))
-			boundValue, _ := strconv.ParseFloat(bound, 64)
-			histogram.Bounds = append(histogram.Bounds, boundValue)
-		} else if strings.HasSuffix(point.Name, "_sum") {
-			histogram.Sum = point.Value
-		} else if strings.HasSuffix(point.Name, "_count") {
-			histogram.Count = uint64(point.Value)
-			histogram.Attributes = attribute.NewSet(metricAttributes...)
-		}
-	}
-
-	s.producer.RecordMetrics(metricdata.Metrics{
-		Name:        de.Name,
-		Description: de.Description,
-		Data: metricdata.Histogram[float64]{
-			DataPoints:  []metricdata.HistogramDataPoint[float64]{histogram},
-			Temporality: metricSdk.DefaultTemporalitySelector(metricSdk.InstrumentKindHistogram),
-		},
-	})
-}
-
-func scrapeEndpoint(target string) []DataEntry {
-	resp, err := http.Get(target)
-	if err != nil {
-		slog.Error("GET to Prometheus HTTP scrape endpoint failed", "endpoint", target)
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		slog.Warn("failed to parse Prometheus scrape endpoint response body")
-	}
-
-	responseString := string(body)
-	splitResponse := strings.Split(responseString, "\n")
-
-	entries := make([]DataEntry, 0)
-
-	currEntry := new(DataEntry)
-	for _, line := range splitResponse {
-		if strings.HasPrefix(line, "# HELP") {
-			// If entry already has content, then we add it to the result slice and reset.
-			if currEntry.Name != "" {
-				entries = append(entries, *currEntry)
-				currEntry = new(DataEntry)
-			}
-			splitHelp := strings.SplitN(line, " ", 4)
-
-			currEntry.Name = splitHelp[2]
-			currEntry.Description = splitHelp[3]
-		} else if strings.HasPrefix(line, "# TYPE") {
-			splitType := strings.SplitN(line, " ", 4)
-			currEntry.Type = splitType[3]
-		} else if line != "" {
-			labels := map[string]string{}
-			splitMetric := strings.SplitN(line, " ", 3)
-			nameAndLabels := splitMetric[0]
-			splitNameAndLabels := strings.Split(nameAndLabels, "{")
-
-			pointName := splitNameAndLabels[0]
-			var pointErr error
-			if len(splitNameAndLabels) > 1 {
-				labelsJson := strings.ReplaceAll("{\""+splitNameAndLabels[1], "=", "\":")
-				labelsJson = strings.ReplaceAll(labelsJson, ",", ",\"")
-				err := json.Unmarshal([]byte(labelsJson), &labels)
-				if err != nil {
-					pointErr = errors.Join(pointErr, fmt.Errorf("failed to parse Prometheus data point labels [%s]: %w", splitNameAndLabels, err))
-				}
-			}
-
-			value, err := strconv.ParseFloat(splitMetric[1], 64)
-			if err != nil {
-				pointErr = errors.Join(pointErr, fmt.Errorf("failed to parse Prometheus data point value [%s]: %w", splitMetric[1], err))
-			}
-
-			if pointErr == nil {
-				currEntry.Values = append(currEntry.Values, DataPoint{
-					Name:   pointName,
-					Labels: labels,
-					Value:  value,
-				})
-			} else {
-				// Discard a data point that has malformed content, so we don't present dirty data to end users.
-				slog.Warn("Prometheus data point parsing error, discarding point", "errors", pointErr)
-			}
-
-		}
-	}
-
-	slog.Debug("Prometheus endpoint scraped", "number of data entries", len(entries), "Prometheus URL", target)
-	return entries
-}
