+++ internal/metrics/exporter_test.go
@@ -4,7 +4,7 @@
  * This source code is licensed under the Apache License, Version 2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-package metric
+package metrics
 
 import (
 	"context"
