+++ internal/metrics/gauge_test.go
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package metric
+package metrics
 
 import (
 	"context"
